refactor(director): 감독 전체조회, 상세조회 코드 수정
